add support for llvm dialect to assembler
