+++ crates/assembler/src/sbpf.pest
@@ -10,8 +10,10 @@ number         = @{ hex_number | decimal_number }
 string_content = @{ (!"\"" ~ ANY)* }
 string_literal = ${ "\"" ~ string_content ~ "\"" }
 
-// Registers (r0 to r10)
-register = @{ "r" ~ ("10" | ASCII_DIGIT) }
+// Registers
+register      = @{ "r" ~ ("10" | ASCII_DIGIT) }
+register_32   = @{ "w" ~ ("10" | ASCII_DIGIT) }
+llvm_register = @{ register | register_32 }
 
 // identifier is used when defining a name
 identifier = @{
@@ -23,8 +25,6 @@ symbol = @{ identifier }
 // Numeric labels ( can be referenced with f or b)
 numeric_label     = @{ ASCII_DIGIT+ }
 numeric_label_ref = @{ numeric_label ~ ("f" | "b") }
-
-label = { (identifier | numeric_label) ~ ":" ~ ((directive_inner | instruction) | NEWLINE) }
 
 // Expressions
 bin_op     = { "+" | "-" | "*" | "/" }
@@ -38,10 +38,13 @@ operand      = { operand_expr | number | symbol }
 // Memory offset.
 memory_op     = { "+" | "-" }
 memory_offset = { number | symbol }
+mem_size      = { "u64" | "u32" | "u16" | "u8" }
 memory_ref    = { "[" ~ register ~ (memory_op ~ memory_offset)+ ~ "]" }
+llvm_memory_ref  = { "(" ~ register ~ (memory_op ~ memory_offset)? ~ ")" }
 
 // Jump target
-jump_target = { numeric_label_ref | (("+" | "-")? ~ number) | symbol }
+signed_number = @{ ("+" | "-")? ~ (hex_number | decimal_number) }
+jump_target   =  { numeric_label_ref | signed_number | symbol }
 
 // ============
 // DIRECTIVES
@@ -88,9 +91,9 @@ directive_inner = {
 }
 directive       = { directive_inner ~ NEWLINE }
 
-// ================
-// INSTRUCTIONS
-// ================
+// ============================
+// DEFAULT INSTRUCTIONS
+// ============================
 
 // Arithmetic and Logical Operations (64-bit)
 alu_64_op       = {
@@ -180,21 +183,60 @@ instr_jump_uncond = { "ja" ~ jump_target }
 endian_op    = { "be16" | "be32" | "be64" | "le16" | "le32" | "le64" }
 instr_endian = { endian_op ~ register }
 
-// Function Call Operations
+// ============================
+// LLVM DIALECT INSTRUCTIONS
+// ============================
+
+// ALU operations
+alu_op           = { "s>>=" | "s/=" | "s%=" | "<<=" | ">>=" | "+=" | "-=" | "*=" | "/=" | "|=" | "&=" | "^=" | "%=" | "=" }
+instr_llvm_alu64 = { register ~ alu_op ~ (register | operand) }
+instr_llvm_alu32 = { register_32 ~ alu_op ~ (register_32 | operand) }
+
+// Negate operations
+instr_llvm_neg64 = { register ~ "=" ~ "-" ~ register }
+instr_llvm_neg32 = { register_32 ~ "=" ~ "-" ~ register_32 }
+
+// Memory load operations
+instr_llvm_load = { llvm_register ~ "=" ~ "*" ~ "(" ~ mem_size ~ "*" ~ ")" ~ llvm_memory_ref }
+instr_llvm_lddw = { register ~ "=" ~ operand ~ "ll" }
+
+// Memory store operation
+instr_llvm_store_reg = { "*" ~ "(" ~ mem_size ~ "*" ~ ")" ~ llvm_memory_ref ~ "=" ~ llvm_register }
+instr_llvm_store_imm = { "*" ~ "(" ~ mem_size ~ "*" ~ ")" ~ llvm_memory_ref ~ "=" ~ operand }
+
+// Byte swap operations
+instr_llvm_endian = { register ~ "=" ~ endian_op ~ register }
+
+// Control flow operations
+cmp_op                 = { "s>=" | "s<=" | "s>" | "s<" | "==" | "!=" | ">=" | "<=" | ">" | "<" | "&" }
+instr_llvm_jump_uncond = { "goto" ~ jump_target }
+instr_llvm_jump_reg    = { "if" ~ register ~ cmp_op ~ register ~ "goto" ~ jump_target }
+instr_llvm_jump_imm    = { "if" ~ register ~ cmp_op ~ operand ~ "goto" ~ jump_target }
+
+// ============================
+// SHARED INSTRUCTIONS
+// ============================
+
 instr_call  = { "call" ~ symbol }
 instr_callx = { "callx" ~ register }
 instr_exit  = { "exit" }
 
-// All instructions
-instruction = {
+// ============================
+// PROGRAM STRUCTURE
+// ============================
+
+instr_default     = {
     (instr_lddw | instr_callx | instr_call | instr_exit | instr_neg32 | instr_neg64 | instr_alu64_imm | instr_alu64_reg | instr_alu32_imm | instr_alu32_reg | instr_load | instr_store_imm | instr_store_reg | instr_jump_imm | instr_jump_reg | instr_jump_uncond | instr_endian) ~ (NEWLINE | &EOI)
 }
+label_default     = { (identifier | numeric_label) ~ ":" ~ ((directive_inner | instr_default) | NEWLINE) }
+statement_default = { NEWLINE | label_default | directive | instr_default }
+program_default   = { SOI ~ statement_default* ~ EOI }
 
-statement = {
-    NEWLINE
-  | label
-  | directive
-  | instruction
+instr_llvm     = {
+    (instr_llvm_lddw | instr_llvm_load | instr_llvm_store_reg | instr_llvm_store_imm | instr_llvm_endian | instr_llvm_neg64 | instr_llvm_neg32 | instr_llvm_alu64 | instr_llvm_alu32 | instr_llvm_jump_uncond | instr_llvm_jump_reg | instr_llvm_jump_imm | instr_callx | instr_call | instr_exit) ~ (NEWLINE | &EOI)
 }
+label_llvm     = { (identifier | numeric_label) ~ ":" ~ ((directive_inner | instr_llvm) | NEWLINE) }
+statement_llvm = { NEWLINE | label_llvm | directive | instr_llvm }
+program_llvm   = { SOI ~ statement_llvm* ~ EOI }
 
-program = { SOI ~ statement* ~ EOI }
+program = _{ program_default | program_llvm }
